+++ src/lib.rs
@@ -50,7 +50,7 @@ target-spec-json = "0.1"
 mod assert_impl;
 
 use process::ProcessBuilder;
-pub use target_spec::{PanicStrategy, TargetArch, TargetEndian, TargetEnv, TargetOs};
+pub use target_spec::{Arch, Env, Os, PanicStrategy, TargetEndian};
 #[path = "gen/target_spec.rs"]
 mod target_spec;
 
@@ -79,7 +79,7 @@ pub struct TargetSpec {
     pub abi: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub allows_weak_linkage: Option<bool>,
-    pub arch: TargetArch,
+    pub arch: Arch,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub archive_format: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -131,8 +131,8 @@ pub struct TargetSpec {
     pub emit_debug_gdb_scripts: Option<bool>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub entry_name: Option<String>,
-    #[serde(default, skip_serializing_if = "TargetEnv::is_none")]
-    pub env: TargetEnv,
+    #[serde(default, skip_serializing_if = "Env::is_none")]
+    pub env: Env,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub executables: Option<bool>,
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -216,8 +216,8 @@ pub struct TargetSpec {
     pub no_builtins: Option<bool>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub no_default_libraries: Option<bool>,
-    #[serde(default, skip_serializing_if = "TargetOs::is_none")]
-    pub os: TargetOs,
+    #[serde(default, skip_serializing_if = "Os::is_none")]
+    pub os: Os,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub obj_is_bitcode: Option<bool>,
     #[serde(skip_serializing_if = "Option::is_none")]